requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
exceptiongroup==1.2.2
flake8==7.1.2
Flask==3.1.0
Flask-Limiter==3.11.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
